+++ docs/changelog.rst
@@ -10,6 +10,7 @@ CHANGELOG
 * Use POI pictures in PDF if the trek has no picture itself
 * Use a placeholder in PDF if there is no picture
 * Parser to import touristic contents from SITRA
+* Add list of all information desks to API
 
 **Bug fixes**